CA2-563: Added Abandonment event reporting.
In so doing, realized it was inappropriate for this report to include contribution
data and removed related code. Contributions cannot be reliably tied to a specific
membership event -- payments link to civicrm_membership, not civicrm_membership_log,
and payments and log events need not occur at the same time. Moreover, memberships
can have more than one contribution associated with them, and the default SQL for
this report joins them such that duplicate rows appear for a single event, which is
nonsensical.